Data Vis update and Score saving update
Added a new colour to the Data Vis so it is easy to read and also
labelled each bar chart with a tag so it is again easier to read. Found
an issue where an incorrect SQL query was being run on all the pages to
update the score. This has been fixed